Populate content and remove external assets
Populated the HTML with dummy content to make the formatting and styling
process easier. Modified the initial styling slightly to account for the
new content containers, but haven't finalised the styling yet.

Initial plan was to use external SVG files and link them into the HTML
to produce more readable HTML code. However, I wanted to play around
with the theme colours of the page. As you can't use CSS to style an
external SVG, I opteded to transition to in-line SVG's for now. Once the
final theme colours are decided, it's possible to re-introduce external
files with the correct colours already applied, link them into the HTML
and remove the styling for the in-line SVG's.